first working select query

+++ tarantool-driver/src/main/java/com/sopovs/moradanen/tarantool/ArrayResult.java
@@ -42,7 +42,7 @@ public class ArrayResult extends AbstractResult<ImmutableArrayValue> {
 	}
 
 	@Override
-	public double getFloat(int index) {
+	public float getFloat(int index) {
 		return current.get(index).asFloatValue().toFloat();
 	}
